Make Day 4 Instructor Discretion
Students encounter LilTwitter at the end of Day 3,
also a day with EE. They don't really attack it
very much on that day and are rather daunted by
the process.

To ease this anxiety (and per their feedback
forms) they wanted and benefitted from a Livecode
of seeing a professional do a slow, thoughtful
migration from a static page into a React app. I
livecoded the Timeline view and Tweets so they
could get a feel of nesting. The general sentiment
was "Oh that's clearer, we just need practice."

I think it's an acceptable outcome.

Also, for this group, their attention had already
shifted to final projects. I think the appropriate
pacing is for day 4 to be a livecode and review
and then after lunch for them to either move to
the data-driven arch (e.g. reddit client) or pivot
to final project preparation. NYC culturally
encourages feasability work on wednesday PM. This
might vary by campus.